Decoupled ndb_api from v1alpha1
