+++ README.md
@@ -222,6 +222,6 @@ Issues and enhancement requests can be submitted in the [Issues tab of this repo
 
 ## License
 
-Copyright 2021-2022 Nutanix, Inc.
+Copyright 2022-2023 Nutanix, Inc.
 
 The project is released under version 2.0 of the [Apache license](http://www.apache.org/licenses/LICENSE-2.0).
