+++ ndb_api/db_helpers.go
@@ -1,0 +1,264 @@
+/*
+Copyright 2022-2023 Nutanix, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package ndb_api
+
+import (
+	"context"
+	"errors"
+	"strconv"
+
+	"github.com/nutanix-cloud-native/ndb-operator/common"
+	"github.com/nutanix-cloud-native/ndb-operator/ndb_client"
+	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
+)
+
+// This function generates and returns a request for provisioning a database (and a dbserver vm) on NDB
+// The database provisioned has a NONE time machine SLA attached to it, and uses the default OOB profiles
+// func GenerateProvisioningRequest(ctx context.Context, ndb_client *ndb_client.NDBClient, dbSpec v1alpha1.DatabaseSpec, reqData map[string]interface{}) (requestBody *DatabaseProvisionRequest, err error) {
+// 	log := ctrllog.FromContext(ctx)
+// 	log.Info("Entered ndb_api_helpers.GenerateProvisioningRequest", "database name", dbSpec.Instance.DatabaseInstanceName, "database type", dbSpec.Instance.Type)
+
+// 	// Fetching the NONE TM SLA
+// 	sla, err := GetNoneTimeMachineSLA(ctx, ndb_client)
+// 	if err != nil {
+// 		log.Error(err, "Error occurred while getting NONE TM SLA", "database name", dbSpec.Instance.DatabaseInstanceName, "database type", dbSpec.Instance.Type)
+// 		return
+// 	}
+
+// 	// Fetch the OOB profiles for the database
+// 	profilesMap, err := GetProfiles(ctx, ndb_client, dbSpec.Instance)
+// 	if err != nil {
+// 		log.Error(err, "Error occurred while getting OOB profiles", "database name", dbSpec.Instance.DatabaseInstanceName, "database type", dbSpec.Instance.Type)
+// 		return
+// 	}
+
+// 	database_names := strings.Join(dbSpec.Instance.DatabaseNames, ",")
+
+// 	// Type assertion
+// 	dbPassword, ok := reqData[common.NDB_PARAM_PASSWORD].(string)
+// 	if !ok || dbPassword == "" {
+// 		err = errors.New("invalid database password")
+// 		var errStatement string
+// 		if !ok {
+// 			errStatement = "Type assertion failed for database password. Expected a string value"
+// 		} else {
+// 			errStatement = "Empty database password"
+// 		}
+// 		log.Error(err, errStatement)
+// 	}
+// 	// Type assertion
+// 	SSHPublicKey, ok := reqData[common.NDB_PARAM_SSH_PUBLIC_KEY].(string)
+// 	if !ok || SSHPublicKey == "" {
+// 		err = errors.New("invalid ssh public key")
+// 		var errStatement string
+// 		if !ok {
+// 			errStatement = "Type assertion failed for SSHPublicKey. Expected a string value"
+// 		} else {
+// 			errStatement = "Empty SSHPublicKey"
+// 		}
+// 		log.Error(err, errStatement)
+// 	}
+// 	// Creating a provisioning request based on the database type
+// 	requestBody = &DatabaseProvisionRequest{
+// 		DatabaseType:             GetDatabaseEngineName(dbSpec.Instance.Type),
+// 		Name:                     dbSpec.Instance.DatabaseInstanceName,
+// 		DatabaseDescription:      "Database provisioned by ndb-operator: " + dbSpec.Instance.DatabaseInstanceName,
+// 		SoftwareProfileId:        profilesMap[common.PROFILE_TYPE_SOFTWARE].Id,
+// 		SoftwareProfileVersionId: profilesMap[common.PROFILE_TYPE_SOFTWARE].LatestVersionId,
+// 		ComputeProfileId:         profilesMap[common.PROFILE_TYPE_COMPUTE].Id,
+// 		NetworkProfileId:         profilesMap[common.PROFILE_TYPE_NETWORK].Id,
+// 		DbParameterProfileId:     profilesMap[common.PROFILE_TYPE_DATABASE_PARAMETER].Id,
+// 		NewDbServerTimeZone:      dbSpec.Instance.TimeZone,
+// 		CreateDbServer:           true,
+// 		NodeCount:                1,
+// 		NxClusterId:              dbSpec.NDB.ClusterId,
+// 		SSHPublicKey:             SSHPublicKey,
+// 		Clustered:                false,
+// 		AutoTuneStagingDrive:     true,
+
+// 		TimeMachineInfo: TimeMachineInfo{
+// 			Name:             dbSpec.Instance.DatabaseInstanceName + "_TM",
+// 			Description:      sla.Description,
+// 			SlaId:            sla.Id,
+// 			Schedule:         make(map[string]string),
+// 			Tags:             make([]string, 0),
+// 			AutoTuneLogDrive: true,
+// 		},
+// 		Nodes: []Node{
+// 			{
+// 				Properties: make([]string, 0),
+// 				VmName:     dbSpec.Instance.DatabaseInstanceName + "_VM",
+// 			},
+// 		},
+// 		ActionArguments: []ActionArgument{
+// 			{
+// 				Name:  "dbserver_description",
+// 				Value: "dbserver for " + dbSpec.Instance.DatabaseInstanceName,
+// 			},
+// 			{
+// 				Name:  "database_names",
+// 				Value: database_names,
+// 			},
+// 			{
+// 				Name:  "db_password",
+// 				Value: dbPassword,
+// 			},
+// 			{
+// 				Name:  "database_size",
+// 				Value: strconv.Itoa(dbSpec.Instance.Size),
+// 			},
+// 		},
+// 	}
+// 	// Setting action arguments based on database type
+// 	dbTypeActionArgs, err := GetActionArgumentsByDatabaseType(dbSpec.Instance.Type)
+
+// 	if err != nil {
+// 		log.Error(err, "Error occurred while getting dbTypeActionArgs", "database type", dbSpec.Instance.Type)
+// 		return
+// 	}
+
+// 	requestBody.ActionArguments = append(requestBody.ActionArguments, dbTypeActionArgs.Get(dbSpec)...)
+
+// 	log.Info("Database Provisioning", "requestBody", requestBody)
+// 	log.Info("Returning from ndb_api_helpers.GenerateProvisioningRequest", "database name", dbSpec.Instance.DatabaseInstanceName, "database type", dbSpec.Instance.Type)
+// 	return
+// }
+
+// This function generates and returns a request for provisioning a database (and a dbserver vm) on NDB
+// The database provisioned has a NONE time machine SLA attached to it, and uses the default OOB profiles
+func GenerateProvisioningRequestt(ctx context.Context, ndb_client *ndb_client.NDBClient, database DatabaseInterface, reqData map[string]interface{}) (requestBody *DatabaseProvisionRequest, err error) {
+	log := ctrllog.FromContext(ctx)
+	log.Info("Entered ndb_api_helpers.GenerateProvisioningRequest", "database name", database.GetDBInstanceName(), "database type", database.GetDBInstanceType())
+
+	// Fetching the NONE TM SLA
+	sla, err := GetNoneTimeMachineSLA(ctx, ndb_client)
+	if err != nil {
+		log.Error(err, "Error occurred while getting NONE TM SLA", "database name", database.GetDBInstanceName(), "database type", database.GetDBInstanceType())
+		return
+	}
+
+	// Fetch the OOB profiles for the database
+	profilesMap, err := GetProfiless(ctx, ndb_client, database.GetDBInstanceType(), database.GetProfileResolvers())
+	if err != nil {
+		log.Error(err, "Error occurred while getting OOB profiles", "database name", database.GetDBInstanceName(), "database type", database.GetDBInstanceType())
+		return
+	}
+
+	database_names := database.GetDBInstanceDatabaseNames()
+
+	// Type assertion
+	dbPassword, ok := reqData[common.NDB_PARAM_PASSWORD].(string)
+	if !ok || dbPassword == "" {
+		err = errors.New("invalid database password")
+		var errStatement string
+		if !ok {
+			errStatement = "Type assertion failed for database password. Expected a string value"
+		} else {
+			errStatement = "Empty database password"
+		}
+		log.Error(err, errStatement)
+	}
+	// Type assertion
+	SSHPublicKey, ok := reqData[common.NDB_PARAM_SSH_PUBLIC_KEY].(string)
+	if !ok || SSHPublicKey == "" {
+		err = errors.New("invalid ssh public key")
+		var errStatement string
+		if !ok {
+			errStatement = "Type assertion failed for SSHPublicKey. Expected a string value"
+		} else {
+			errStatement = "Empty SSHPublicKey"
+		}
+		log.Error(err, errStatement)
+	}
+	// Creating a provisioning request based on the database type
+	requestBody = &DatabaseProvisionRequest{
+		DatabaseType:             GetDatabaseEngineName(database.GetDBInstanceType()),
+		Name:                     database.GetDBInstanceName(),
+		DatabaseDescription:      "Database provisioned by ndb-operator: " + database.GetDBInstanceName(),
+		SoftwareProfileId:        profilesMap[common.PROFILE_TYPE_SOFTWARE].Id,
+		SoftwareProfileVersionId: profilesMap[common.PROFILE_TYPE_SOFTWARE].LatestVersionId,
+		ComputeProfileId:         profilesMap[common.PROFILE_TYPE_COMPUTE].Id,
+		NetworkProfileId:         profilesMap[common.PROFILE_TYPE_NETWORK].Id,
+		DbParameterProfileId:     profilesMap[common.PROFILE_TYPE_DATABASE_PARAMETER].Id,
+		NewDbServerTimeZone:      database.GetDBInstanceTimeZone(),
+		CreateDbServer:           true,
+		NodeCount:                1,
+		NxClusterId:              database.GetNDBClusterId(),
+		SSHPublicKey:             SSHPublicKey,
+		Clustered:                false,
+		AutoTuneStagingDrive:     true,
+
+		TimeMachineInfo: TimeMachineInfo{
+			Name:             database.GetDBInstanceName() + "_TM",
+			Description:      sla.Description,
+			SlaId:            sla.Id,
+			Schedule:         make(map[string]string),
+			Tags:             make([]string, 0),
+			AutoTuneLogDrive: true,
+		},
+		Nodes: []Node{
+			{
+				Properties: make([]string, 0),
+				VmName:     database.GetDBInstanceName() + "_VM",
+			},
+		},
+		ActionArguments: []ActionArgument{
+			{
+				Name:  "dbserver_description",
+				Value: "dbserver for " + database.GetDBInstanceName(),
+			},
+			{
+				Name:  "database_names",
+				Value: database_names,
+			},
+			{
+				Name:  "db_password",
+				Value: dbPassword,
+			},
+			{
+				Name:  "database_size",
+				Value: strconv.Itoa(database.GetDBInstanceSize()),
+			},
+		},
+	}
+	// Setting action arguments based on database type
+	// dbTypeActionArgs, err := GetActionArgumentsByDatabaseType(database.GetDBInstanceType())
+
+	if err != nil {
+		log.Error(err, "Error occurred while getting dbTypeActionArgs", "database type", database.GetDBInstanceType())
+		return
+	}
+
+	requestBody.ActionArguments = append(requestBody.ActionArguments, database.GetDBInstanceActionArguments()...)
+
+	log.Info("Database Provisioning", "requestBody", requestBody)
+	log.Info("Returning from ndb_api_helpers.GenerateProvisioningRequest", "database name", database.GetDBInstanceName(), "database type", database.GetDBInstanceType())
+	return
+}
+
+// Returns a request to delete a database instance
+func GenerateDeprovisionDatabaseRequest() (req *DatabaseDeprovisionRequest) {
+	req = &DatabaseDeprovisionRequest{
+		Delete:               true,
+		Remove:               false,
+		SoftRemove:           false,
+		Forced:               false,
+		DeleteTimeMachine:    true,
+		DeleteLogicalCluster: true,
+	}
+	return
+}
