+++ ndb_api/interfaces.go
@@ -1,3 +1,19 @@
+/*
+Copyright 2022-2023 Nutanix, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package ndb_api
 
 import (
@@ -8,8 +24,23 @@ import (
 
 type ProfileResolver interface {
 	Resolve(ctx context.Context, allProfiles []ProfileResponse, filter func(p ProfileResponse) bool) (profile ProfileResponse, err error)
+	GetName() string
+	GetId() string
 }
 
 type DatabaseActionArgs interface {
 	Get(dbSpec v1alpha1.DatabaseSpec) []ActionArgument
+}
+
+type ProfileResolvers map[string]ProfileResolver
+
+type DatabaseInterface interface {
+	GetDBInstanceName() string
+	GetDBInstanceType() string
+	GetDBInstanceDatabaseNames() string
+	GetDBInstanceTimeZone() string
+	GetDBInstanceSize() int
+	GetNDBClusterId() string
+	GetDBInstanceActionArguments() []ActionArgument
+	GetProfileResolvers() ProfileResolvers
 }
